requirements.txt
numpy
pandas
flask
pathlib
dataclasses
mysql-connector-python
pymysql
python-dotenv
scikit-learn
# -e .